+++ Flask Sample/requirements.txt
@@ -1,0 +1,2 @@
+Flask == 0.12.2
+translate
