--- requirements.txt
@@ -1,2 +1,0 @@
-Flask == 0.12.2
-translate
